requirements.txt
py-cpuinfo
psutil
vk_api
